+++ packages/core/styles/src/packages/atomic/text/nimbus-text.types.ts
@@ -3,6 +3,9 @@ import {
   TextAlign,
   TextDecoration,
   WordBreak,
+  TextOverflow,
+  WhiteSpace,
+  Overflow,
 } from "../../../types";
 import { textSprinkle } from "./nimbus-text.css";
 
@@ -62,4 +65,17 @@ export interface TextSprinkle {
    * The textDecoration shorthand property sets the appearance of decorative lines on text.
    */
   textDecoration?: TextDecoration | TextConditions<TextDecoration>;
+  /**
+   * The textOverflow property specifies how overflowed content that is not displayed should be signaled to the user.
+   * Note: This property only works when overflow is set to 'hidden'.
+   */
+  textOverflow?: TextOverflow | TextConditions<TextOverflow>;
+  /**
+   * The whiteSpace property sets how white space inside an element is handled.
+   */
+  whiteSpace?: WhiteSpace | TextConditions<WhiteSpace>;
+  /**
+   * The overflow property specifies what happens if content overflows an element's box.
+   */
+  overflow?: Overflow | TextConditions<Overflow>;
 }